Refactor: admin 컴포넌트 구조 대폭 개선 및 타입 시스템 통합 - admin 컴포넌트 폴더 구조 평면화 (하위 폴더 제거) - 모든 컴포넌트를 admin 폴더 바로 아래로 이동 - 컴포넌트 내부 interface를 types/components/admin.ts로 분리 - API 타입들을 types/api/common.ts로 통합 (Post, PostFormData 등) - UI 설정을 types/ui/admin.ts로 분리 (categoryInfo) - pages/admin/data 폴더 완전 제거 - 모든 import 경로를 @/types로 통일
