--- src/components/admin/CreatePost/index.ts
@@ -1,1 +1,0 @@
-export { CreatePostForm } from './CreatePostForm';
