--- src/components/admin/Dashboard/index.ts
@@ -1,3 +1,0 @@
-export { DashboardHeader } from './DashboardHeader';
-export { DashboardStats } from './DashboardStats';
-export { DashboardTable } from './DashboardTable';
