--- src/components/admin/EditPost/index.ts
@@ -1,1 +1,0 @@
-export { EditPostForm } from './EditPostForm';
