--- src/components/admin/Login/index.ts
@@ -1,1 +1,0 @@
-export { LoginForm } from './LoginForm';
