+++ src/components/admin/index.ts
@@ -1,4 +1,6 @@
-export * from './Login';
-export * from './Dashboard';
-export * from './CreatePost';
-export * from './EditPost';
+export * from './LoginForm';
+export * from './DashboardHeader';
+export * from './DashboardStats';
+export * from './DashboardTable';
+export * from './CreatePostForm';
+export * from './EditPostForm';
