+++ src/pages/admin/EditPost.tsx
@@ -1,19 +1,17 @@
 import { useState, useEffect } from 'react';
 import { useNavigate, useParams, useSearchParams } from 'react-router-dom';
-// import { useAuth } from '@/contexts/AuthContext';
 import { useApi } from '@/hooks/useApi';
 import { postsApi } from '@/api/admin/posts';
 import type { PostDetailResponse } from '@/types/api';
-import { EditPostForm } from '@/components/admin/EditPost';
+import { EditPostForm } from '@/components/admin';
 import { Button } from '@/components/ui/button';
-import type { PostFormData, PostCategory, Post } from './data';
-import { categoryInfo } from './data';
+import type { PostFormData, PostCategory, Post } from '@/types/api/common';
+import { categoryInfo } from '@/types/ui/admin';
 
 export default function EditPost() {
   const navigate = useNavigate();
   const { id } = useParams<{ id: string }>();
   const [searchParams] = useSearchParams();
-  // const { user } = useAuth(); // 사용되지 않음
   const [formData, setFormData] = useState<PostFormData>({
     title: '',
     content: '',
