--- src/pages/admin/data/index.ts
@@ -1,1 +1,0 @@
-export * from './postTypes';
